Move BufferedStreamWriter (+Switchable, +Pipe) to RESPite.Streams as the single copy
The trio existed twice: here (redis-coupled factory) and drifted lib-generic copies on the proxy spike
branch. Duplicated code is asking for trouble, so this is the unification, replace/move not duplicate:
the SE.Redis files are GONE, the one copy lives in RESPite.Streams (it already leaned on RESPite.Buffers
-- the move is with the grain), and the proxy branch deletes its copies on next rebase.

What decoupled: the factory loses ConnectionType/ConfigurationOptions and takes (WriteMode, Stream,
MemoryPool<byte>?, CancellationToken); WriteMode.Default maps to Async as the lib-safe default. The
redis POLICY (pub/sub never wants sync latency mode) moves to the one call site in
PhysicalConnection.InitOutput, where it reads as policy instead of hiding in a lib factory.

RESPite gains the same conditional System.IO.Pipelines reference SE.Redis already carries (in-box from
net10.0) so PipeStreamWriter -- the comparison/troubleshooting implementation -- stays with its family
rather than being split back out via cross-assembly inheritance. Everything stays internal; RESPite
already grants IVT to StackExchange.Redis and the test projects.

Solution builds 0 errors on all TFMs; BufferedStreamWriter + RoundTrip + InProcess/WriteMode test
batteries all pass (129 tests).
