+++ tests/StackExchange.Redis.Tests/RoundTripUnitTests/TestConnection.cs
@@ -4,6 +4,7 @@ using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using RESPite.Streams;
 using RESPite.Tests;
 using Xunit;
